Delete the CRD definition when the operator is stopped
When the operator is stopped and started, the CRD definition stays the same. There is a problem with this: changes to the CRD definition won't take effect if the definition is not updated to the API server. Deleting the definition when the operator stops and recreating it when the operator starts effectively updates the definition.