Add an entrypoint to spark-operator for handling non-root users
In order to use the spark-on-k8s-operator on a platform where a user
is assigned based on a security context (OpenShift), the /etc/passwd
file needs to be modified in an entrypoint like it is in the standard
Apache Spark 2.3 image to support an assigned user. Also, to avoid overriding
the entrypoint, the command in spark-operator.yaml should be left defaulted
but the args still passed. The /etc/passwd file will be unchanged for a root
user since root is already in the file.