Create builds with spark:2.4.5-SNAPSHOT
This patch set builds the operator using the spark image:
- gcr.io/spark-operator/spark:2.4.5-SNAPSHOT

This is needed as documented in this issue:
- https://github.com/GoogleCloudPlatform/spark-on-k8s-operator/issues/669

The underlying problem is that openjdk on alpine has become unsupported
and the spark project has moved to openjdk:8-jdk-slim which is ubuntu
based as described in this issue:
- https://issues.apache.org/jira/browse/SPARK-28938

Without this, the GKE service accounts or workload identity won't
properly authorize.  This was not observed with the given examples
because they explicitly pass in credentials rather than relying on
workload identity or service accounts.