fix: Remove priority from the spec when using priority class
By default pods created by kube will have a priority of 0 set in their
spec, or the value of the PriorityClass acting as a `globalDefault`.
When using the `batchSchedulerOptions.priorityClassName` the spec was
mutated by adding the `priorityClassName` to the spec, but the
`priority` was left unchanged.

This caused an error "the integer value of priority (0) must not be
provided in pod spec; priority admission controller computed X from
the given PriorityClass name." where X is the priority associated
to the PrioritClass kube object.

The patcher now properly remove the existing priority when a
priorityClassName is applied, letting kube apply the correct priority to
the executor and driver pods.