Split values' rbac.create flag in createRole and createClusterRole
This is the second part of changes allowing blue/green deployments of
the Spark Operator: one can now deploy the ClusterRole without the Role,
which would be defined in the common part of a blue/green deployment.

Note this is a breaking change, so the chart version has been bumped to 2.