Improve github CI, and helm default docker version
Add a docker build into the github CI, even if we don't use it, and also
stop using latest as the default container version, as it is clearly wrong, and
the latest tag hasn't been updated in a long time.

travis-ci.org has stopped building the code for this repository as well, so we should
focus on github workflows. Ideally we'd move the building of the docker container into
github as well because then at least all CI is managed in one place.

Signed-off-by: Tom Hellier <me@tomhellier.com>